% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapsisConnection.R
\name{createC4ScriptInstance}
\alias{createC4ScriptInstance}
\title{Create an instance of C4Script for a particular model}
\usage{
createC4ScriptInstance(modelName)
}
\arguments{
\item{modelName}{the anme of the model (e.g. artemis)}
}
\description{
Create an instance of C4Script for a particular model
}
