% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapsisConnection.R
\docType{data}
\name{rcapsisSettingEnv}
\alias{rcapsisSettingEnv}
\title{The settings environment for this package}
\format{
An object of class \code{environment} of length 0.
}
\usage{
rcapsisSettingEnv
}
\description{
This environment contains the general settings of the package.
}
\keyword{datasets}
