% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCapsisConnection.R
\name{setCapsisPath}
\alias{setCapsisPath}
\title{Set the path to the Capsis Platform}
\usage{
setCapsisPath(path)
}
\arguments{
\item{path}{the complete path to Capsis as in the example below. The file.path function should be used to
define the path}
}
\description{
Set the path to the Capsis Platform
}
\examples{
myPath <- file.path("C:","MyFolder","Capsis")
# setCapsisPath(myPath)  ### not run

}
\seealso{
file.path
}
